requirements.txt
flask
flask-dance
python-dotenv
oauthlib
PyPDF2
python-docx
groq
blinker
cryptography
httpx
